fix: harmony: Correctly separate this counters when sharing Max
